delete VideoCompositor.jsx; rename Compositor to VideoCompositor
